bugfix: submitting a completed bite results in a big crash
When attempting to submit a bite, if a NoSuchElementException
occurs while looking for the buttons to click to pull the user's
code from GitHub we will stop processing, inform the user that
they've already finished this bite and then open the bite page
for confirmation.

There also was bug for opening the page with a bare uninitialized
format string.