feature: added support for archive file caching
The download command now looks for bite archive files in a directory
named "cache" found in the current directory (if present).  If the
archive file is found, the download from the platform is skipped and
the archive is extracted in the current directory.  If the archive is
not found, it's downloaded from the platform and saved to the "cache"
directory before extracting. The `--keep-zip` boolean flag argument to
`download` is now `--cleanup` and will cause the downloaded and cached
archive file to be removed automatically (opposite of the previous
flags usage).